+++ pkg/processors/blobber/impl_read.go
@@ -49,25 +49,6 @@ func getBLOBKeyRead(ctx context.Context, work pipeline.IWorkpiece) (err error) {
 	return nil
 }
 
-func getBLOBKeyWrite(ctx context.Context, work pipeline.IWorkpiece) (err error) {
-	bw := work.(*blobWorkpiece)
-	if bw.isPersistent() {
-		bw.blobKey = &iblobstorage.PersistentBLOBKeyType{
-			ClusterAppID: istructs.ClusterAppID_sys_blobber,
-			WSID:         bw.blobMessageWrite.wsid,
-			BlobID:       bw.newBLOBID,
-		}
-	} else {
-		// temp
-		bw.blobKey = &iblobstorage.TempBLOBKeyType{
-			ClusterAppID: istructs.ClusterAppID_sys_blobber,
-			WSID:         bw.blobMessageWrite.wsid,
-			SUUID:        bw.newSUUID,
-		}
-	}
-	return nil
-}
-
 func initResponse(ctx context.Context, work pipeline.IWorkpiece) (err error) {
 	bw := work.(*blobWorkpiece)
 	bw.writer = bw.blobMessageRead.okResponseIniter(
