+++ pkg/sys/sys.vsql
@@ -60,7 +60,7 @@ ABSTRACT WORKSPACE Workspace (
 		OwnerQName2 text
 	) WITH Tags=(WorkspaceOwnerTableTag);
 
-	TABLE BLOB INHERITS WDoc (status int32 NOT NULL);
+	TABLE BLOB INHERITS WDoc (status int32 NOT NULL) WITH Tags=(WorkspaceOwnerTableTag);
 
 	TABLE Subject INHERITS sys.CDoc (
 		Login varchar NOT NULL,
